cmd/k8s-operator,k8s-operator: add ProxyGroup CRD
The ProxyGroup CRD specifies a set of N pods which will each be a
tailnet device, and will have M different ingress or egress services
mapped onto them. It is the mechanism for specifying how highly
available proxies need to be. This commit only adds the definition, no
controller loop, and so it is not currently functional.

Updates #13406

Signed-off-by: Tom Proctor <tomhjp@users.noreply.github.com>