cmd/k8s-operator,k8s-operator,kube: add a reconciler for ExternalName Services for egress HA
Adds a new reconciler that reconciles ExternalName Services that define a
tailnet target that should be exposed to cluster workloads on a ProxyGroup's
proxies.
The reconciler ensures that for each such service, the config mounted to
the proxies is updated with the tailnet target definition and that
and EndpointSlice and ClusterIP Service are created for the service.

Signed-off-by: Irbe Krumina <irbe@tailscale.com>