logtail: remove LowMemory mode
Fixes tailscale/tailscale#13685

logtail currently has a LowMemory flag, which is enabled upon initialization on memory-constrained platforms like iOS. This flag reduces the ring buffer size from 256 log lines to 64. It was introduced over four years ago, back when Tailscale supported iOS 14 and earlier, where network extensions were limited to 15 MB of RAM.

Since the memory limit has now increased to 50 MB on all supported iOS versions (with our minimum requirement being iOS 15.0), the need to conserve a few kilobytes of RAM by reducing buffer entries and the size of each flush is minimal.

The additional code paths are more things we need to maintain over time... with little benefit. This PR removes that.

To be merged after we cut the first 1.77 unstable, as this might be a risky change.

Signed-off-by: Andrea Gottardo <andrea@gottardo.me>