derp: extend the sclient write deadline to avoid possible i/o timeouts
We see some probers suffering i/o timeouts that do not have a clear
explanation. One working theory is that we're scheduling writes and
setting the deadline too long before the actual write occurs, and that
is causing us to timeout. Extend the deadline at the flush point, so as
to attempt to avoid this in cases where a lot of data has been buffered,
which seems like the most likely point we could have spent too long
doing work before writing.

Updates tailscale/corp#23668

Signed-off-by: James Tucker <james@tailscale.com>