ipn/ipnlocal: clear app connector storage on 'down'
App connectors store their routes in the state file, so that they may
have an understanding of which routes can be removed when app connectors
are reconfigured.

App connectors route storage is deleted if the AdvertisedRoutes are
edited by something else, such as `tailscale set`.

This should also happen on `tailscale down`.

Fixes #13680
Signed-off-by: Fran Bull <fran@tailscale.com>