net/sockstats: prevent crash in setNetMon
Fixes tailscale/tailscale#13983

As @raggi pointed out, DefaultRouteInterface and Interface are set at different points in time. The map lookup in `setNetMon` should check whether the map contains a value for the interface name before getting the interface index. If it doesn't, it returns a zero value containing a pointer which will be nil in the zero value case, causing the panic.

Signed-off-by: Andrea Gottardo <andrea@gottardo.me>