cmd/k8s-operator: auto sign node keys in locked tailnet
This adds support for running the k8s operator in a locked tailnet. When the operator is running in a locked tailnet, it will sign node keys for devices which it manages. This is done by adding a new reconciler to the operator, `node-key-signer`, which will sign node keys for managed devices using the operator's own key.

The `node-key-signer` reconciler is enabled by setting the `OPERATOR_AUTO_SIGN_NODE_KEYS` environment variable to `true` (which is the default).

In order for the operator to sign node keys, the operator's public lock key must be trusted by the tailnet. If it's not, a warning will be logged and the operator will not sign node keys.

If the tailnet is not locked, no operation will be performed.

Fixes #13273

Signed-off-by: Jake Scaltreto <jake@scaltreto.com>